Rename ValueRange to domain and move domain entries up one level
